feat: 와플카드, LikeToggle, EditBox 타입스크립트로 전환
