+++ .eslintrc
@@ -15,7 +15,8 @@
       "plugins": ["@typescript-eslint", "prettier"],
       "extends": ["plugin:@typescript-eslint/recommended", "prettier"],
       "rules": {
-        "@typescript-eslint/explicit-module-boundary-types": "off"
+        "@typescript-eslint/explicit-module-boundary-types": "off",
+        "@typescript-eslint/no-explicit-any": "off"
       }
     }
   ],
