+++ src/components/atom/BackButton/BackButton.stories.jsx
@@ -1,4 +1,4 @@
-import { BackButton } from '@/components';
+import BackButton from './BackButton';
 
 export default {
   title: 'Component/Base/BackButton',
