+++ src/components/index.ts
@@ -2,7 +2,6 @@ export { default as Button } from './atom/Button/Button';
 export { default as ColorPicker } from './molecule/ColorPicker/ColorPicker';
 export { default as Modal } from './molecule/Modal/Modal';
 export { default as Text } from './atom/Text/Text';
-export { default as Card } from './molecule/Card/Card';
 export { default as Icons } from './atom/Icons';
 export { default as Portal } from './atom/Portal/Portal';
 export { default as Spinner } from './molecule/Spinner/Spinner';
@@ -11,6 +10,7 @@ export { default as Input } from './atom/Input/Input';
 export { default as ScrollGuide } from './molecule/ScrollGuide/ScrollGuide';
 export { default as EditBox } from './molecule/EditBox/EditBox';
 export { default as WaffleCard } from './organism/WaffleCard/WaffleCard';
+export { default as EmptyCard } from './organism/WaffleCard/EmptyCard';
 export { default as CardEditModal } from './organism/CardEditModal/CardEditModal';
 export { default as ChattingCardModal } from './organism/ChattingCardModal/ChattingCardModal';
 export { default as Tab } from './molecule/Tab/Tab';
