+++ src/components/molecule/LikeToggle/LikeToggle.tsx
@@ -7,17 +7,25 @@ import FavoriteBorderRoundedIcon from '@mui/icons-material/FavoriteBorderRounded
 import Common from '@/styles';
 import { Text } from '@/components';
 
+interface LikeToggleProps extends Omit<React.ComponentProps<'div'>, 'onClick'> {
+  toggled?: boolean;
+  interactive?: boolean;
+  onClick?: (toggled: boolean, count: number) => void;
+  count?: number;
+}
+
 const LikeToggle = ({
   toggled,
   interactive,
   onClick,
   count: initCount = 0,
   ...props
-}) => {
+}: LikeToggleProps) => {
   const [toggle, setToggle] = useState(toggled);
   const [count, setCount] = useState(initCount);
 
-  const handleClick = e => {
+  // TODO(윤호): any타입 삭제하고 명확한 타입 지정하는 방법 찾아보기
+  const handleClick = (e: any) => {
     e.stopPropagation();
 
     if (!interactive) return;
