+++ src/components/organism/WaffleCard/EmptyCard.style.ts
@@ -1,10 +1,12 @@
-import React from 'react';
-import PropTypes from 'prop-types';
 import styled from '@emotion/styled';
-import { Text } from '@/components';
 import Common from '@/styles';
+import { Text as RawText } from '@/components';
 
-const Container = styled.div`
+interface ContainerProps extends React.ButtonHTMLAttributes<HTMLDivElement> {
+  backgroundColor?: string;
+}
+
+export const Container = styled.div<ContainerProps>`
   display: flex;
   flex-direction: column;
   justify-content: center;
@@ -38,35 +40,6 @@ const Container = styled.div`
   }
 `;
 
-const IconText = styled(Text)`
+export const Text = styled(RawText)`
   margin-bottom: 6px;
 `;
-
-const Empty = ({ fontSize, iconSize, onClick, ...props }) => {
-  const handleClick = e => {
-    onClick && onClick(e);
-  };
-
-  return (
-    <Container onClick={handleClick} {...props}>
-      <IconText size={'75px'} color={Common.colors.secondary}>
-        +
-      </IconText>
-      <Text size={fontSize} color={Common.colors.secondary}>
-        카드 만들기
-      </Text>
-    </Container>
-  );
-};
-
-Empty.propTypes = {
-  fontSize: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
-  iconSize: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
-  onClick: PropTypes.func,
-};
-
-Empty.defaultProps = {
-  iconSize: '4rem',
-};
-
-export default Empty;
