+++ src/types/index.ts
@@ -1,0 +1,2 @@
+export type { UserType } from './user';
+export type { WaffleCardType } from './waffleCard';
