+++ src/types/user.ts
@@ -1,0 +1,5 @@
+export interface UserType {
+  id: string;
+  name: string;
+  email: string;
+}
